Update project to use MS VC++ v14
Add MLST command (RFC 3659)
Add OpenSSL INI options; OpenSSL_SecurityLevel, OpenSSL_LoadLegacyProvider, OpenSSL_Ciphers13 (TLS 1.3 cipher suite)
Added Version.h to centralize version numbering
Added TLS 1.3 support with automatic best TLS version negotiation
Added OpenSSL 3.6.1 re-write and included support for older OpenSSL compatibility
Updated TCL 8.5 to 9.0.2
Update AUTH SSL handing to be synonym of AUTH TLS (SSL removed)
Update Minimum Windows version to Vista
Fix MessageWindow handler
Fix Socket usage to not crash, lock, have thread issues or linger
Fix string file path comparison to not crash on non-ascii chars
Fix buffer overrun and precision issues
Fix XCRC flag hiding it from features
See change log

+++ include/Client.h
@@ -33,7 +33,7 @@ typedef struct _CLIENTSLOT
 typedef struct _CLIENT
 {
 	ONLINEDATA		Static;
-	DWORD           dwTransferLastUpdated; // didn't want to change exported ONLINEDATA field
+	ULONGLONG       dwTransferLastUpdated; // didn't want to change exported ONLINEDATA field
 	LPIOSERVICE		lpService;
 	DWORD			dwLoginTime;
